rename .collector.CollectorUtil to .util.IOUtils
